Create 16x16 grid with color-in
This commit represents the first three steps of the project: creating a
webpage with a 16x16 grid of square divs and setting up a "hover" effect
so that the grid divs change color when the mouse passes over them. The
dimensions of the container are hard-coded to 960px x 960px, but the
squares are dynamically sized using percentages and flex-basis. An event
listener is added to each square during the creation loop to change the
background color when the mouse enter event is triggered. The color
change is persistent, like a real etch-a-sketch.