Add button to clear and redraw the screen
This commit represents the remaining steps for the most basic version of
the project. A button is added to the top of the page which when clicked
prompts the user for a numerical input, then the draw screen is cleared
(by calling .replaceChildren()) and redrawn according to the number of
squares (defined by the user input).