Change cursor to only paint while mouse down
This commit refactors the script a bit to allow for the cursor to only
paint the canvas while the mouse button is down, rather than by simply
hovering over the square. Two event listeners are added to the document
root to toggle a boolean variable isMouseDown based on each mouseDown
and mouseUp event.

The redrawScreen() function no longer appends
an event listener during div creation, but rather uses a
querySelectorAll().forEach to grab the pixels after creation and add two
event listeners. The first event fires whenever the mouse enters the
div, checking if isMouseDown is currently true; the second merely fires
when the mousedown targets the div. In both cases, the pixel.classList
adds "painted" to paint the square.