Add draw and erase modes
This commit introduces the ability to select between "draw mode" and
"erase mode" by clicking the appropriate buttons in the UI. There is now
a global variable called currentTool, an "enum" (frozen object) that
maintains the possible tools (draw, erase, shade, lighten), a dictionary
called ToolHandlers with the respective functions that each tool should
call, and a function called applyTool which validates the handler first
and then calls the function with the pixel passed as argument. This is a
big leap that allows me to partition out the responsibility of the code
and organize things in a way that is extensible and readable. In that
same vein, the script is reorganized into sections (delineated by
comments) since things were starting to get unwieldy.