Implement shading and lightening features
This commit takes advantage of the RGB and HSL logic introduced in the
last commit to allow for shading and lightening of pixels. In brief, the
tools when called on a pixel obtain the current background color, parse
the color string to an RGB object, convert that to HSL, increment or
decrement the lightness by 10%, and then chain it backwards to set a new
background color on the pixel. This may be more verbose than simply
layering overlays divs or affecting the opacity value, but it was fun to
manage the math and learn how to pass around objects.

These functions also made me rethink the nature of an unpainted pixel.
Rather than "" or "transparent", I decided to go ahead and establish
each pixels background color as rbg(255, 255, 255) during creation in
the call to redrawScreen(). This allows me to avoid safety checking
during the parseColorString() function since I've already managed the
pixel's state from the get-go.