Revise method for changing canvas dimensions
With this commit, the user now adjusts the pixel dimensions of the
canvas (drawScreen) by inputting a value into an input element and
pressing enter or click the resizeButton. The event listener parses the
input for an integer, falling back to a default value of 16 if the value
is falsy or NaN. Then as usual, clearScreen() is called and the
userInput is passed as argument to a call to redrawScreen(). This last
call handles clamping (2-64), and also updates the value of the input to
any clamped value to reflect the change.