Add color selection feature
This commit updates the interface so that the user can now use a
color-picker input to change the current color of the brush. The browser
is pretty much in control of the UI behavior for this, but I did style
the input and label to fit in with the rest of the site.